Changed app.py and template files
Accounted for variables such as age, sex and height in the daily protein count via a bmr calculation. Sets the protein count to 1 decimal place and added an optional box for the user to fill in the amount of protein for the food if they know the exact amount. Website requires you to fill in profile page before they can reach the index page and start entering foods (so the protein count can be calculated).